requirements.txt
tqdm
requests
bs4
psutil
sqlalchemy
pyyaml

cffi==1.15.1
clingo==5.6.2
pycparser==2.21
git+https://github.com/yuce/pyswip@master#egg=pyswip
tqdm==4.66.1
mysql-connector-python
SQLAlchemy
pytest
pytest-cov
coverage
python-sat
clingo
spacy
pandas
reasonable
duckdb
numpy
psutil
networkx
git+https://github.com/logic-and-learning-lab/Popper@main
bitarray
oracledb
mysql-connector-python
pytest-asyncio
pytest-mock
colorama